Extracted all Exception classes to their own files, and removed the last "#if SFX".
